Convert legacy node:test to Vitest for the issue-729 wheel guard
